Move levenshtein distance function to yt/funcs.py
